+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Unreleased
 * Improved `Delta` support: added `Delta` model and two new methods; `since` and `longpoll`.
+* Fix Virtual Calendar logic and serialization
 
 ### 6.1.0 / 2022-01-28
 * Add support for `Event` to ICS